mobile: handle removing urbit client on logout
